+++ src/main/java/com/en_circle/slt/plugin/ui/console/SltREPL.java
@@ -1,6 +1,7 @@
 package com.en_circle.slt.plugin.ui.console;
 
 import com.en_circle.slt.plugin.ui.PackageSelectorComponent;
+import com.en_circle.slt.plugin.ui.SltCoreWindow;
 import com.intellij.openapi.project.Project;
 import com.intellij.ui.tabs.TabInfo;
 
@@ -8,8 +9,17 @@ import java.awt.*;
 
 public class SltREPL extends SltConsole {
 
-    public SltREPL(Project project) {
+    private final SltCoreWindow window;
+
+    public SltREPL(Project project, SltCoreWindow window) {
         super(project);
+        this.window = window;
+    }
+
+    @Override
+    public void close() {
+        window.removeComponent(this);
+        super.close();
     }
 
     @Override