+++ src/main/java/com/en_circle/slt/plugin/SltStatusBarWidgetFactory.java
@@ -1,0 +1,43 @@
+package com.en_circle.slt.plugin;
+
+import com.en_circle.slt.plugin.ui.SltPackageWidget;
+import com.intellij.openapi.project.Project;
+import com.intellij.openapi.wm.StatusBar;
+import com.intellij.openapi.wm.StatusBarWidget;
+import com.intellij.openapi.wm.StatusBarWidgetFactory;
+import org.jetbrains.annotations.Nls;
+import org.jetbrains.annotations.NonNls;
+import org.jetbrains.annotations.NotNull;
+
+public class SltStatusBarWidgetFactory implements StatusBarWidgetFactory {
+
+    @Override
+    public @NonNls @NotNull String getId() {
+        return "Common Lisp";
+    }
+
+    @Override
+    public @Nls @NotNull String getDisplayName() {
+        return SltBundle.message("slt.ui.statusbar.name");
+    }
+
+    @Override
+    public boolean isAvailable(@NotNull Project project) {
+        return true;
+    }
+
+    @Override
+    public @NotNull StatusBarWidget createWidget(@NotNull Project project) {
+        return new SltPackageWidget(project);
+    }
+
+    @Override
+    public void disposeWidget(@NotNull StatusBarWidget widget) {
+
+    }
+
+    @Override
+    public boolean canBeEnabledOn(@NotNull StatusBar statusBar) {
+        return true;
+    }
+}