+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 ### Fixes
 
 - Fixed action threading
+- Fixed repl not closing
 
 ## 0.2.1 - 230120