+++ CHANGELOG.md
@@ -2,8 +2,18 @@
 
 ### Added
 
+- Current package at editor cursor widget
+
 ### Fixes
 
+- Fixed action threading
+- Fixed repl not closing
+
+## 0.2.1 - 230120
+
+### Fixes
+
+- Attempting to fix windows path error
 
 ## 0.2.0 - 230120